+++ sideline/components/providers/ReturnSessionTracker.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { tryEmitReturnSession } from "@/lib/productAnalytics";
+import { useEffect } from "react";
+
+export function ReturnSessionTracker() {
+  useEffect(() => {
+    tryEmitReturnSession();
+  }, []);
+  return null;
+}